+++ app/scripts/types.ts
@@ -1,5 +1,7 @@
 /* eslint-disable */
 
+import type { THEME_DARK, THEME_LIGHT } from "./configs";
+
 export type Scale = import("d3-scale").ScaleContinuousNumeric<number, number>;
 
 export type TrackPosition = typeof import('./configs/primitives').TRACK_LOCATIONS[number];
@@ -77,3 +79,27 @@ export interface TrackObject {
   movedY(extent: number): void;
   zoomedY(yPosition: number, wheelDelta: number): void;
 }
+
+export type Theme = typeof THEME_DARK | typeof THEME_LIGHT;
+
+type TilesetInfoBase = {
+  name: string;
+  coordSystem: string;
+  min_pos: number[];
+  max_pos: number[];
+  max_zoom: number;
+  tile_size?: number;
+  max_tile_width?: number;
+  transforms?: { name: string, value: string }[];
+}
+
+export type LegacyTilesetInfo = TilesetInfoBase & {
+  max_width: number;
+  bins_per_dimension?: number;
+}
+
+export type ResolutionsTilesetInfo = TilesetInfoBase & {
+  resolutions: number[];
+}
+
+export type TilesetInfo = LegacyTilesetInfo | ResolutionsTilesetInfo;